components/corpusgrabber/requirements.txt
langgraph
langchain-openai
langchain-anthropic
langchain_community
matplotlib
networkx
mypy-protobuf
pytest
jsonschema
pika
sqlalchemy
psycopg2-binary
pyyaml
tabulate
selenium
colorama
magika
redis
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp